requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==25.1
numpy==2.2.5
python-dotenv==1.0.0
Pillow
django-allauth==0.58.2
openai==1.3.5
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7 
psycopg2-binary